fix: Auto reject terminated Snap dialogs (#15051)
<!--
Please submit this PR as a draft initially.
Do not mark it as "Ready for review" until the template has been
completely filled out, and PR status checks have passed at least once.
-->

## **Description**

If a Snap crashes we want to reject it's dialog so that it closes since
the dialog will be unresponsive anyway. This mirrors the implementation
in the extension:
https://github.com/MetaMask/metamask-extension/blob/7a22dbc2ea9d47efd4901c7326ccf3ab4218b1c3/app/scripts/metamask-controller.js#L3113-L3130