chore: Bump Snaps packages (#15477)
<!--
Please submit this PR as a draft initially.
Do not mark it as "Ready for review" until the template has been
completely filled out, and PR status checks have passed at least once.
-->

## **Description**

This bumps Snaps packages to the latest versions. Notable changes are:

- Support for `wallet_switchEthereumChain` inside Snaps.
- Like with websites, we want to support a custom network per Snap
rather than relying on the globally selected network. For Snaps, the
permission for the network to switch to is automatically granted.
- Support for SIP-31 `onClientRequest` handler.
- New `showTestnets` property in `snap_getPreferences`.
- Added `isMinimumPlatformVersion` action.

For `wallet_switchEthereumChain` we have applied the same changes as on
the extension where `isAddNetworkFlow` was changed to be `autoApprove`.
It seems like `autoApprove` was already attempted to be used.

## **Manual testing steps**

### `wallet_switchEthereumChain`

1. Go to
[`test-snaps`](https://metamask.github.io/snaps/test-snaps/latest/)
2. Install the Ethereum Provider Example Snap.
3. Switch the chain ID using the `test-snaps` selector for the Ethereum
Provider example.
4. Click "Get version".
5. The Snap should now use the new network, without any permission
requests, and the appropriate chain ID should be shown.

---------

Co-authored-by: Maarten Zuidhoorn <maarten@zuidhoorn.com>