fix: Improve Snaps UI scrolling experience (#16426)
<!--
Please submit this PR as a draft initially.
Do not mark it as "Ready for review" until the template has been
completely filled out, and PR status checks have passed at least once.
-->

## **Description**

Wraps Snap content in a Touchable component as a hack to enable proper
scrolling. Before this the scrolling would be inconsistent and work only
when tapping certain UI elements.

The actual diff here is ~20 lines changed, but due to changing the
wrapping it affected all snapshots. I recommend reviewing with
whitespace off to reduce the diff massively.

## **Related issues**

Fixes: https://github.com/MetaMask/metamask-mobile/issues/16376

## **Manual testing steps**

1. Go to a new in-app browser tab
2. Navigate to magiceden.io (ensure you're in a disconnected state - log
out from the dapp and remove permissions for the dapp in the wallet as
well)
3. Find the MetaMask wallet connection option
5. Connect through the Solana interface
6. After connecting a Solana account, choose "Click to sign" on the
Magic Eden
7. Scroll through the message