fix: cp-7.50.0 Prevent cronjob state from getting out of sync (#16711)
<!--
Please submit this PR as a draft initially.
Do not mark it as "Ready for review" until the template has been
completely filled out, and PR status checks have passed at least once.
-->

## **Description**

Prevent the `CronjobController` state from getting out of sync by
initializing the controller before the `SnapController` allowing it to
subscribe to the controller events. Additionally introduce a `init`
function that handles the side-effects previously present in the
`CronjobController` constructor.

Since this is a release blocker, this upstream change was patched in:
https://github.com/MetaMask/snaps/commit/a9f0277aba928d9980f7ff827b1879535f2dda4b

## **Related issues**

Fixes: https://github.com/MetaMask/metamask-mobile/issues/16700

## **Manual testing steps**

1. Downgrade Solana Snap to 1.30.3
2. Create a fresh MetaMask instance
3. Upgrade Solana Snap to 1.30.4 and reload the app
4. Wait a couple minutes
5. Notice that there are no errors in the console