fix: cp-7.50.0 Mitigate flash of safe withdrawal limit message (#16727)
<!--
Please submit this PR as a draft initially.
Do not mark it as "Ready for review" until the template has been
completely filled out, and PR status checks have passed at least once.
-->

## **Description**

This PR mitigates the flash of withdrawal limit message that appears on
withdrawal forms for lending.

The improvement is to set the withdrawal limit value to undefined and do
normal flow if it never becomes defined. If there is a failure, we
fallback and do not stop the user from withdrawing as the default 0
withdrawal amount would be incorrect in this case.

<!--
Write a short description of the changes included in this pull request,
also include relevant motivation and context. Have in mind the following
questions:
1. What is the reason for the change?
2. What is the improvement/solution?
-->

## **Related issues**

Fixes: https://consensyssoftware.atlassian.net/browse/TAT-1059
Fixes: https://github.com/MetaMask/metamask-mobile/issues/16749

## **Manual testing steps**

- Go to bsc chain with popular networks off (to avoid a diff bug)
- Attempt to withdraw, you should not see any flash of message
- You should be able to withdraw (if bsc rpc is not faulty) 
- If you cannot withdraw restart the app and try again to see if bsc
issue is cleared
- Try linea withdraw as well it should work the same and have no flash
of message

## **Screenshots/Recordings**

<!-- If applicable, add screenshots and/or recordings to visualize the
before and after of your change. -->

### **Before**

<!-- [screenshots/recordings] -->

### **After**

<!-- [screenshots/recordings] -->

## **Pre-merge author checklist**

- [x] I’ve followed [MetaMask Contributor
Docs](https://github.com/MetaMask/contributor-docs) and [MetaMask Mobile
Coding
Standards](https://github.com/MetaMask/metamask-mobile/blob/main/.github/guidelines/CODING_GUIDELINES.md).
- [x] I've completed the PR template to the best of my ability
- [x] I’ve included tests if applicable
- [ ] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [x] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-mobile/blob/main/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.

## **Pre-merge reviewer checklist**

- [x] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [x] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.